refactor: rename `theme-toggle.tsx` into `ThemeToggle.tsx`
