+++ components/Navbar.tsx
@@ -5,8 +5,8 @@ import { cn } from '@/lib/utils'
 import { ExternalLinkIcon, HandHeartIcon, MenuIcon } from 'lucide-react'
 import Link from 'next/link'
 import * as React from 'react'
+import { ThemeToggle } from './ThemeToggle'
 import GitHubIcon from './icons/GitHubIcon'
-import { ThemeToggle } from './theme-toggle'
 import { Badge } from './ui/badge'
 import { Button } from './ui/button'
 import {
